requirements.txt
tensorflow==2.18.0
pandas
numpy
opencv-python
scikit-learn
matplotlib
streamlit
pillow
